Refact: Common RecordView 제거, CalendarRecordView 생성
